+++ src/index.js
@@ -1,8 +1,8 @@
 // Browser-compatible utilities (pure JS)
 export {default as Collection} from "./browser/lib/Collection.js"
 export {default as Data} from "./browser/lib/Data.js"
-export {default as Disposable} from "./browser/lib/Disposable.js"
-export {Disposable as DisposableClass} from "./browser/lib/Disposable.js"
+export {default as Disposer} from "./browser/lib/Disposer.js"
+export {Disposer as DisposerClass} from "./browser/lib/Disposer.js"
 export {default as Type} from "./browser/lib/TypeSpec.js"
 export {default as Valid} from "./lib/Valid.js"
