+++ tests/node/Disposable.test.js
@@ -3,14 +3,14 @@
 import assert from "node:assert/strict"
 import {describe, it} from "node:test"
 
-import {DisposableClass as Disposable} from "@gesslar/toolkit"
+import {DisposerClass as Disposer} from "@gesslar/toolkit"
 
-describe("Disposable (node entry)", () => {
+describe("Disposer (node entry)", () => {
   it("is exported and disposes registered callbacks", () => {
     const calls = []
-    const disposable = new Disposable()
+    const disposable = new Disposer()
 
-    disposable.registerDisposer(() => calls.push("done"))
+    disposable.register(() => calls.push("done"))
     disposable.dispose()
 
     assert.deepEqual(calls, ["done"])
